DEV: Isolate precession-nutation functionality.
